+++ web/packages/new/package.json
@@ -2,6 +2,10 @@
     "name": "@/new",
     "version": "0.0.0",
     "private": true,
-    "dependencies": {},
+    "dependencies": {
+        "@/next": "*",
+        "@/utils": "*",
+        "@ente/shared": "*"
+    },
     "devDependencies": {}
 }
